+++ packages/hs-react-ui/src/index.ts
@@ -3,14 +3,16 @@ import Card from './components/Card';
 import Checkbox from './components/Checkbox';
 import Divider from './components/Divider';
 import Dropdown from './components/Dropdown';
+import InteractionFeedback from './components/InteractionFeedback';
 import Label from './components/Label';
 import Modal from './components/Modal';
 import RangeSlider from './components/RangeSlider';
 import Table from './components/Table';
+import Tag from './components/Tag';
 import Text from './components/Text';
 import TextInput from './components/TextInput';
 import Progress from './components/Progress';
-import { FoundryProvider, FoundryContext, useColors } from './context';
+import { FoundryProvider, FoundryContext, useTheme } from './context';
 
 import colors from './enums/colors';
 import timings from './enums/timings';
@@ -23,16 +25,18 @@ export {
   Checkbox,
   Divider,
   Dropdown,
+  InteractionFeedback,
   Label,
   Modal,
   RangeSlider,
   Table,
+  Tag,
   Text,
   TextInput,
   Progress,
   FoundryProvider,
   FoundryContext,
-  useColors,
+  useTheme,
   colors,
   timings,
   fonts,